requirements.txt
google-adk>=1.32.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.126.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12
fastmcp>=2.0.0
starlette>=0.41.0
itsdangerous>=2.2.0